+++ src/draw_world.c
@@ -6,7 +6,7 @@
 /*   By: cgosseli <cgosseli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/21 11:38:39 by jeepark           #+#    #+#             */
-/*   Updated: 2022/10/04 14:57:33 by cgosseli         ###   ########.fr       */
+/*   Updated: 2022/10/04 17:38:56 by cgosseli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 in the world. Put the image to the window and launch the mlx. */
 void    draw_world(t_world *world, t_mlx *mlx)
 {
-	int		y;
-	int		x;
+	double		y;
+	double		x;
 	t_ray	ray;
 	int		obj_idx;
 	t_viewport		viewport;
@@ -26,16 +26,16 @@ void    draw_world(t_world *world, t_mlx *mlx)
 	
 	viewport = get_viewport(world->cam);
 	// get_background_color(world->ambient_light);
-    while (y < SIZEY)
+    while (y < (double)SIZEY)
 	{
 		x = 0;
-		while (x <= SIZEX)
+		while (x <= (double)SIZEX)
 		{
 			ray = set_ray(world->cam, x, y, viewport);
 			obj_idx = hit_obj(&ray, world);
 			if (obj_idx >= 0)
 			{
-				put_pix(mlx, x, y, compute_color(&ray, world->objects[obj_idx], world));
+				put_pix(mlx, (int)x, (int)y, compute_color(&ray, world->objects[obj_idx], world));
 			}
 			x++;
 		}
